pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "openvla"
authors = [
    {name = "Moo Jin Kim", email="moojink@stanford.edu"},
    {name = "Karl Pertsch", email="pertsch@berkeley.edu"},
    {name = "Siddharth Karamcheti", email="skaramcheti@cs.stanford.edu"},
]
description = "OpenVLA: Vision-Language-Action Models for Robotics"
version = "0.0.3"
readme = "README.md"
requires-python = ">=3.8"
keywords = ["vision-language-actions models", "multimodal pretraining", "robot learning"]
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "accelerate",
    "draccus",
    "einops",
    #"flash_attn==2.5.5",      # Here for documentation -- install *AFTER* editable install (follow README)
    "huggingface_hub",
    "json-numpy",
    "jsonlines",
    "matplotlib",
    "peft",
    "protobuf",
    "rich",
    "sentencepiece",
    "timm==0.9.10",
    "tokenizers",     # Test it!!!
    #"torch==2.2.0",
    #"torchvision==0.17.0",
    #"torchaudio==2.2.0",
    "transformers==4.48.2",   # Test it!!!
    "wandb",
    "tensorflow",
    "tensorflow_datasets",
    "tensorflow_graphics",
    "dlimp @ git+https://github.com/moojink/dlimp_openvla",
    "pyzmq",
    "vllm==0.7.2",
    "ray==2.40.0",
    "cupy-cuda12x",
    "pandas",
    "termcolor",
    "tensorboard",
]

[project.optional-dependencies]
dev = [
    "black>=24.2.0",
    "gpustat",
    "ipython",
    "pre-commit",
    "ruff>=0.2.2",
]
sagemaker = [
    "boto3",
    "sagemaker"
]

[project.urls]
homepage = "https://github.com/openvla/openvla"
repository = "https://github.com/openvla/openvla"
documentation = "https://github.com/openvla/openvla"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["cache"]

[tool.setuptools.package-data]
"prismatic" = ["py.typed"]

[tool.black]
line-length = 121
target-version = ["py38", "py39", "py310"]
preview = true

[tool.ruff]
line-length = 121
target-version = "py38"

[tool.ruff.lint]
select = ["A", "B", "E", "F", "I", "RUF", "W"]
ignore = ["F722"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]
